+++ run/pg_duckdb_analytics/models/marts/tpch_daily_revenue.sql
@@ -15,7 +15,7 @@
     total_quantity numeric(38,2),
     revenue numeric(38,2),
     revenue_after_tax numeric(38,2),
-    avg_discount numeric
+    avg_discount double precision
     
     )
  ;
@@ -38,7 +38,7 @@ select
     round(
         sum(lineitems.extended_price * (1 - lineitems.discount) * (1 + lineitems.tax)), 2
     )::numeric as revenue_after_tax,
-    avg(lineitems.discount)::numeric as avg_discount
+    avg(lineitems.discount) as avg_discount
 from "analytics"."dev"."stg_lineitems" as lineitems
