+++ run/pg_duckdb_analytics/models/marts/tpch_q1_pricing_summary.sql
@@ -14,9 +14,9 @@
     sum_base_price numeric,
     sum_disc_price numeric,
     sum_charge numeric,
-    avg_qty numeric,
-    avg_price numeric,
-    avg_disc numeric,
+    avg_qty double precision,
+    avg_price double precision,
+    avg_disc double precision,
     count_order bigint
     
     )
@@ -36,9 +36,9 @@
     round(sum(lineitems.extended_price), 2) as sum_base_price,
     round(sum(lineitems.extended_price * (1 - lineitems.discount)), 2) as sum_disc_price,
     round(sum(lineitems.extended_price * (1 - lineitems.discount) * (1 + lineitems.tax)), 2) as sum_charge,
-    avg(lineitems.quantity)::numeric as avg_qty,
-    avg(lineitems.extended_price)::numeric as avg_price,
-    avg(lineitems.discount)::numeric as avg_disc,
+    avg(lineitems.quantity) as avg_qty,
+    avg(lineitems.extended_price) as avg_price,
+    avg(lineitems.discount) as avg_disc,
     count(*) as count_order
 from "analytics"."dev"."stg_lineitems" as lineitems
 where lineitems.ship_date <= date '1998-09-02'
