+++ run/pg_duckdb_analytics/models/marts/tpch_supplier_performance.sql
@@ -14,7 +14,7 @@
     region_name text,
     revenue numeric,
     order_count bigint,
-    avg_discount numeric
+    avg_discount double precision
     
     )
  ;
